requirements.txt
SQLAlchemy==1.4.42
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.1
frozenlist==1.3.1
future==0.18.2
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
line-bot-sdk==2.3.0
multidict==6.0.2
pydantic==1.10.2
pytest==7.1.3
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
yarl==1.8.1
